+++ src/pybel/struct/mutation/__init__.py
@@ -2,11 +2,12 @@
 
 """This module contains functions that mutate or make transformations on a network."""
 
-from . import collapse, deletion, expansion, induction, inference, metadata, transfer, utils
+from . import collapse, deletion, expansion, induction, induction_expansion, inference, metadata, transfer, utils
 from .collapse import *
 from .deletion import *
 from .expansion import *
 from .induction import *
+from .induction_expansion import *
 from .inference import *
 from .metadata import *
 from .transfer import *
@@ -17,6 +18,7 @@ __all__ = (
         deletion.__all__ +
         expansion.__all__ +
         induction.__all__ +
+        induction_expansion.__all__ +
         inference.__all__ +
         metadata.__all__ +
         transfer.__all__ +
