+++ src/pybel/struct/mutation/induction/__init__.py
@@ -1,10 +1,18 @@
 # -*- coding: utf-8 -*-
 
-from . import upstream, utils
+from . import annotations, citation, neighborhood, paths, upstream, utils
+from .annotations import *
+from .citation import *
+from .neighborhood import *
+from .paths import *
 from .upstream import *
 from .utils import *
 
 __all__ = (
+        annotations.__all__ +
+        citation.__all__ +
+        neighborhood.__all__ +
+        paths.__all__ +
         upstream.__all__ +
         utils.__all__
 )
